ConDEc-507: Move the summarizaionEventListener
